finished report, added comments for source

+++ pcb.c
@@ -12,6 +12,9 @@
 #include "pcb.h"
 
 struct pcb *startpcb(struct pcb *process) {
+  if (process == NULL) {
+    return NULL;
+  }
   pid_t pid = 0;
   switch (pid = fork()) {
     case -1:
@@ -22,12 +25,13 @@ struct pcb *startpcb(struct pcb *process) {
       printf("hostd: error executing process\n");
       return NULL;
   }
+  // retrieve the fields for printing
   int a = process->priority;
   int b = process->remainingcputime;
   int c = process->procmem->offset;
   int d = process->procmem->size;
   int e,f,g,h;
-  if (a == 0) {
+  if (a == 0) { // realtime processes have priority 0 and no resources
     e = 0;
     f = 0;
     g = 0;
@@ -39,13 +43,16 @@ struct pcb *startpcb(struct pcb *process) {
     g = process->resources->rsrcs[MODEM];
     h = process->resources->rsrcs[CD];
   }
-  printf("pid\tpty\trem\toffset\tsize\tprinter\tscanner\tmodem\tcd\n");
-  printf("%d\t%d\t%d\t%d\t%d\t%d\t%d\t%d\t%d\n",pid,a,b,c,d,e,f,g,h);
+  printf("pid\tprty\trem\toffset\tsize\tprinter\tscanner\tmodem\tcd\tstatus\n");
+  printf("%d\t%d\t%d\t%d\t%d\t%d\t%d\t%d\t%d\tRUNNING\n",pid,a,b,c,d,e,f,g,h);
   process->pid = pid;
   return process;
 }
 
 struct pcb *terminatepcb(struct pcb *process) {
+  if (process == NULL) {
+    return NULL;
+  }
   int status;
   kill(process->pid, SIGINT);
   waitpid(process->pid, &status, WUNTRACED);
@@ -54,6 +61,9 @@ struct pcb *terminatepcb(struct pcb *process) {
 }
 
 struct pcb *suspendpcb(struct pcb *process) {
+  if (process == NULL) {
+    return NULL;
+  }
   int status;
   kill(process->pid, SIGTSTP);
   waitpid(process->pid, &status, WUNTRACED);
@@ -62,6 +72,9 @@ struct pcb *suspendpcb(struct pcb *process) {
 }
 
 struct pcb *restartpcb(struct pcb *process) {
+  if (process == NULL) {
+    return NULL;
+  }
   if (kill(process->pid, SIGCONT)) {
     return NULL;
   }
@@ -80,10 +93,15 @@ struct pcb *createnullpcb(void) {
 }
 
 struct pcb *enqpcb(struct pcb **head, struct pcb *process) {
+  if (process == NULL || head == NULL) {
+    return NULL;
+  }
   if (*head == NULL) {
     *head = process; // if the head is empty, set head to be the process
     return *head;
   }
+  // else, traverse the linked list until we come to the end
+  // then we queue the given process onto the end
   struct pcb *curr = *head;
   while (curr->next) {
     curr = curr->next;
@@ -93,22 +111,30 @@ struct pcb *enqpcb(struct pcb **head, struct pcb *process) {
 }
 
 struct pcb *deqpcb(struct pcb **head) {
-  if (*head == NULL) {
+  if (head == NULL) {
     return NULL;
   }
+  if (*head == NULL) { // if there's nothing at the head, we can't dequeue anything
+    return NULL;
+  }
+  // else, remove the current head and update the head pointer
   struct pcb *old = *head;
   struct pcb *new = (*head)->next;
   *head = new;
   old->next = NULL;
-  return old;
+  return old; // return the process we just removed
 }
 
 void printq(struct pcb **head) {
-  if (!*head) {
+  if (head == NULL) {
+    return;
+  }
+  if (!*head) { // there's nothing at the head, so the list is empty
     return;
   }
   struct pcb *curr = *head;
-  while (curr->next) {
+  while (curr->next) { 
+  // traverse the linked list, printing out various process fields for debugging
     printf("ID: %d, arr time: %d, mbytes: %d\n", curr->pid, curr->arrivaltime, curr->mbytes);
     curr = curr->next;
   }
